rate limiting scoped to middleware instead of request
