+++ src/LimitsMiddleware/InterlockedBooleanExtensions.cs
@@ -1,0 +1,10 @@
+﻿namespace LimitsMiddleware
+{
+    internal static class InterlockedBooleanExtensions
+    {
+        public static bool EnsureCalledOnce(this InterlockedBoolean interlockedBoolean)
+        {
+            return interlockedBoolean.CompareExchange(true, false);
+        }
+    }
+}
