+++ src/LimitsMiddleware/LimitsMiddleware.csproj
@@ -46,6 +46,8 @@
       <Link>Properties\SharedAssemblyInfo.cs</Link>
     </Compile>
     <Compile Include="App_Packages\LibOwin.1.0\LibOwin.cs" />
+    <Compile Include="InterlockedBoolean.cs" />
+    <Compile Include="InterlockedBooleanExtensions.cs" />
     <Compile Include="Limits.cs" />
     <Compile Include="ConnectionTimeoutOptions.cs" />
     <Compile Include="ContentLengthExceededException.cs" />
@@ -60,6 +62,7 @@
     <Compile Include="OptionsBase.cs" />
     <Compile Include="Properties\Annotations.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RateLimiter.cs" />
     <Compile Include="StringExtension.cs" />
     <Compile Include="ThrottledStream.cs" />
     <Compile Include="TimeoutStream.cs" />
